+++ src/Annotation/PropertyWrapperTrait.php
@@ -1,0 +1,21 @@
+<?php
+
+namespace Kassko\DataMapper\Annotation;
+
+/**
+* Common code for property wrappers
+*
+* @author kko
+*/
+trait PropertyWrapperTrait
+{
+    /**
+    * @var string
+    */
+    public $prefix;
+
+    /**
+    * @var string
+    */
+    public $name;
+}