+++ src/ClassMetadataLoader/InnerYamlLoader.php
@@ -3,20 +3,22 @@
 namespace Kassko\DataMapper\ClassMetadataLoader;
 
 use Kassko\DataMapper\ClassMetadata\ClassMetadata;
+use Kassko\DataMapper\Configuration\Configuration;
+use Symfony\Component\Yaml\Parser;
 
 /**
- * Class metadata loader for php format.
+ * Class metadata loader for yaml embeded in objects.
  *
  * @author kko
  */
-class PhpLoader extends AbstractLoader
+class InnerYamlLoader extends AbstractLoader
 {
     private $classMetadata;
 
     public function supports(LoadingCriteriaInterface $loadingCriteria)
     {
         return
-            'php' === $loadingCriteria->getResourceType()
+            'inner_yaml' === $loadingCriteria->getResourceType()
             &&
             method_exists($loadingCriteria->getResourceClass(), $loadingCriteria->getResourceMethod())
         ;
@@ -25,7 +27,7 @@ class PhpLoader extends AbstractLoader
     protected function doGetData(LoadingCriteriaInterface $loadingCriteria)
     {
         $callable = [$loadingCriteria->getResourceClass(), $loadingCriteria->getResourceMethod()];
-        return $callable();
+        return $this->parseContent($callable());
     }
 
     protected function doLoadClassMetadata(ClassMetadata $classMetadata, array $data)
@@ -40,16 +42,16 @@ class PhpLoader extends AbstractLoader
 
     private function loadClassAnnotations(array $data)
     {
-        if (isset($data['object']['repositoryClass'])) {
-            $this->classMetadata->setRepositoryClass($data['object']['repositoryClass']);
+        if (isset($data['object']['providerClass'])) {
+            $this->classMetadata->setRepositoryClass($data['object']['providerClass']);
         }
 
-        if (isset($data['object']['readDateFormat'])) {
-            $this->classMetadata->setObjectReadDateFormat($data['object']['readDateFormat']);
+        if (isset($data['object']['readDateConverter'])) {
+            $this->classMetadata->setObjectReadDateFormat($data['object']['readDateConverter']);
         }
 
-        if (isset($data['object']['writeDateFormat'])) {
-            $this->classMetadata->setObjectWriteDateFormat($data['object']['writeDateFormat']);
+        if (isset($data['object']['writeDateConverter'])) {
+            $this->classMetadata->setObjectWriteDateFormat($data['object']['writeDateConverter']);
         }
 
         if (isset($data['object']['propertyAccessStrategy'])) {
@@ -131,10 +133,6 @@ class PhpLoader extends AbstractLoader
         $dataName = 'fields';
         foreach ($data[$dataName] as $mappedFieldName => $fieldData) {
 
-            if (is_numeric($mappedFieldName)) {//if $mappedFieldName is a numeric index, $fieldData contains the field.
-                $fieldData = ['name' => $fieldData];
-            }
-
             $mappedManagedFieldNames[] = $mappedFieldName;
 
             if (! isset($fieldData['name'])) {
@@ -163,7 +161,7 @@ class PhpLoader extends AbstractLoader
                $mappedDateFieldNames[] = $mappedFieldName;
             }
 
-            if (isset($fieldData['writeStrategy']) || isset($fieldData['readStrategy'])) {
+            if (isset($fieldData['writeConverter']) || isset($fieldData['readConverter'])) {
 
                 $fieldsWithHydrationStrategy[$mappedFieldName] = [];
                 $fieldsWithHydrationStrategy[$mappedFieldName][ClassMetadata::INDEX_EXTRACTION_STRATEGY] = null;
@@ -171,12 +169,12 @@ class PhpLoader extends AbstractLoader
                 $fieldsWithHydrationStrategy[$mappedFieldName][ClassMetadata::INDEX_EXTENSION_CLASS] = null;
             }
 
-            if (isset($fieldData['writeStrategy'])) {
-                $fieldsWithHydrationStrategy[$mappedFieldName][ClassMetadata::INDEX_EXTRACTION_STRATEGY] = $fieldData['writeStrategy'];
+            if (isset($fieldData['writeConverter'])) {
+                $fieldsWithHydrationStrategy[$mappedFieldName][ClassMetadata::INDEX_EXTRACTION_STRATEGY] = $fieldData['writeConverter'];
             }
 
-            if (isset($fieldData['readStrategy'])) {
-                $fieldsWithHydrationStrategy[$mappedFieldName][ClassMetadata::INDEX_HYDRATION_STRATEGY] = $fieldData['readStrategy'];
+            if (isset($fieldData['readConverter'])) {
+                $fieldsWithHydrationStrategy[$mappedFieldName][ClassMetadata::INDEX_HYDRATION_STRATEGY] = $fieldData['readConverter'];
             }
 
             if (isset($fieldData['mappingExtensionClass'])) {
@@ -271,5 +269,10 @@ class PhpLoader extends AbstractLoader
         if (count($fieldsWithHydrationStrategy)) {
             $this->classMetadata->setFieldsWithHydrationStrategy($fieldsWithHydrationStrategy);
         }
+    }
+
+    protected function parseContent($content)
+    {
+        return (new Parser())->parse($content);
     }
 }